Fix ultrasound pulse measurement overflowing at long distance This occurs above ~5.6m

+++ src/src.ino
@@ -172,7 +172,7 @@ void processCommand() {
 			digitalWrite(pulse, LOW);
 
 			// measure the echo time on the echo pin
-			int duration = pulseIn(echo, HIGH, 60000);
+			long duration = pulseIn(echo, HIGH, 60000);
 			Serial.print(microsecondsToMm(duration));
 			Serial.print("\n");
 			return;
